Enable bugprone-implicit-widening-of-multiplication-result clang-tidy check (#421)

+++ tasks/seq/example/perf_tests/main.cpp
@@ -13,7 +13,7 @@
 class NesterovTaskSeqTest : public ::testing::TestWithParam<ppc::core::PerfResults::TypeOfRunning> {
  protected:
   static void RunTest(ppc::core::PerfResults::TypeOfRunning mode) {
-    constexpr int kCount = 500;
+    constexpr size_t kCount = 500;
 
     // Create data
     std::vector<int> in(kCount * kCount, 0);
